fix: refactor to be compatible with existing structure
