+++ packages/networks/gno-dev/index.ts
@@ -11,7 +11,6 @@ export const gnoDevNetwork: GnoNetworkInfo = {
     NetworkFeature.SocialFeed,
     NetworkFeature.UPP,
     NetworkFeature.GnoProjectManager,
-    NetworkFeature.NameService,
   ],
   featureObjects: [
     {
