+++ packages/screens/UserPublicProfile/components/UPPIntro.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useState } from "react";
+import React, { useState } from "react";
 import { Linking, Platform, useWindowDimensions, View } from "react-native";
 
 import { PremiumSubscriptionModal } from "./modals/PremiumSubscriptionModal";
@@ -26,12 +26,10 @@ import { usePremiumIsSubscribed } from "@/hooks/feed/usePremiumIsSubscribed";
 import { useMaxResolution } from "@/hooks/useMaxResolution";
 import { useNSUserInfo } from "@/hooks/useNSUserInfo";
 import useSelectedWallet from "@/hooks/useSelectedWallet";
-import { useUserProfile } from "@/hooks/useUserProfile";
 import {
   accountExplorerLink,
   getNetworkFeature,
   NetworkFeature,
-  NetworkKind,
   parseUserId,
 } from "@/networks";
 import { DEFAULT_NAME } from "@/utils/social-feed";
@@ -53,8 +51,6 @@ export const UPPIntro: React.FC<{
 }> = ({ userId, isUserOwner, setIsEditProfileModal = (val) => {} }) => {
   const selectedWallet = useSelectedWallet();
   const { metadata } = useNSUserInfo(userId);
-  const { data: profile, isLoading: isLoadingProfile } = useUserProfile(userId);
-
   const { copyToClipboard } = useCopyToClipboard();
   const socialButtonStyle = { margin: layout.spacing_x0_75 };
   const [network, userAddress] = parseUserId(userId);
@@ -75,28 +71,6 @@ export const UPPIntro: React.FC<{
     useState(false);
   const [premiumSubscriptionModalVisible, setPremiumSubscriptionModalVisible] =
     useState(false);
-
-  const { bio, displayName } = useMemo(() => {
-    if (isLoadingProfile) return {};
-
-    let bio = metadata.public_bio;
-    let displayName = metadata.public_name;
-
-    if (network?.kind === NetworkKind.Gno) {
-      bio = profile?.bio;
-      displayName = profile?.displayName;
-    }
-    return { bio, displayName };
-  }, [
-    network?.kind,
-    profile?.bio,
-    profile?.displayName,
-    metadata.public_bio,
-    metadata.public_name,
-    isLoadingProfile,
-  ]);
-
-  const username = metadata.tokenId;
 
   return (
     <>
@@ -302,13 +276,15 @@ export const UPPIntro: React.FC<{
       >
         <View>
           {/* Pseudo and bio */}
-          {username ? (
+          {metadata.tokenId ? (
             <>
-              <BrandText style={[fontBold16]}>{displayName}</BrandText>
+              <BrandText style={[fontBold16]}>
+                {metadata?.public_name}
+              </BrandText>
               <BrandText
                 style={[fontMedium14, { color: neutral55, marginTop: 2 }]}
               >
-                @{username}
+                @{metadata.tokenId}
               </BrandText>
             </>
           ) : (
@@ -327,7 +303,7 @@ export const UPPIntro: React.FC<{
               { maxWidth: 735, marginTop: layout.spacing_x1 },
             ]}
           >
-            {bio}
+            {metadata.public_bio}
           </BrandText>
         </View>
         {/* Stats and public address */}
